Changes after presentation

+++ API/Models/AmbrosiaEntity.cs
@@ -14,8 +14,8 @@ namespace Ambrosia.Models
             this.RowKey = IMEI;
         }
         public AmbrosiaEntity() { }
-        public float XCoord { get; set; }
-        public float YCoord { get; set; }
+        public double XCoord { get; set; }
+        public double YCoord { get; set; }
         
 
     }
